+++ migrations/scripts/stock_portfolio.sql
@@ -4,7 +4,7 @@ CREATE TABLE stock_portfolio(
     shares_count INT,
     collectible_id INT,
     user_id INT,
-    total_price FLOAT,
+    total_price NUMERIC(15, 2),
     transaction_id INT,
     FOREIGN KEY (collectible_id) REFERENCES users(user_id),
     FOREIGN KEY (user_id) REFERENCES users(user_id),
